test(omo-senpi): add active task TUI scenario
